# SimBlog17API/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SimBlog17API",
    "version": "1.0.0",
    "description": "Lightweight forum backend API.\nSupports JWT access/refresh tokens, articles, comments, interactions, homepage aggregation, and admin moderation.\n"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local"
    }
  ],
...
  "paths": {
...
    "/api/comments": {
      "post": {
        "tags": [
          "Comment"
        ],
        "summary": "Create comment or reply",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStatusResponseCommentResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          }
        }
      }
    },
...
    "/api/comments/{id}": {
      "delete": {
        "tags": [
          "Comment"
        ],
        "summary": "Delete comment",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStatusResponseCommentResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
    "/api/admin/users/{userId}/promote": {
      "put": {
        "tags": [
          "Admin"
        ],
        "summary": "Promote user to admin",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStatusResponseUserAccessResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/admin/users/{userId}/demote": {
      "put": {
        "tags": [
          "Admin"
        ],
        "summary": "Demote admin to user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStatusResponseUserAccessResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/admin/articles/pending": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List pending articles",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/Size"
          },
          {
            "$ref": "#/components/parameters/Sort"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStatusResponsePageArticleMetaResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/admin/articles/{id}/status": {
      "put": {
        "tags": [
          "Admin"
        ],
        "summary": "Update article status",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ArticleIdPath"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminArticleStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStatusResponseArticleMetaResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          }
        }
      }
    },
    "/api/admin/comments": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List comments by status",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/Size"
          },
          {
            "$ref": "#/components/parameters/Sort"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStatusResponsePageCommentResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/admin/comments/{id}": {
      "delete": {
        "tags": [
          "Admin"
        ],
        "summary": "Delete comment as admin",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStatusResponseCommentResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/admin/comments/{id}/status": {
      "put": {
        "tags": [
          "Admin"
        ],
        "summary": "Update comment moderation status",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminCommentStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStatusResponseCommentResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          }
        }
      }
    },
    "/api/admin/media/cleanup": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Cleanup orphan media files",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "olderThanDays",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 7,
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStatusResponseLong"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "InvalidRequest": {
        "description": "Invalid Request - validation or payload format error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiStatusResponseString"
            },
            "example": {
              "statusCode": 2,
              "statusMessage": "Invalid Request",
              "timeStamp": 1778003000000,
              "traceId": "f84f2d4d-7f13-4bd1-8b63-ec2f58b1f54b",
              "data": "Validation failed: username must not be blank"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized - missing/invalid credentials",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiStatusBase"
            },
            "example": {
              "statusCode": 5,
              "statusMessage": "Unauthorized",
              "timeStamp": 1778002000000,
              "traceId": "5f8a6ca2-b8bf-4ad4-b3d4-0d7f7dd6a20b"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Forbidden - insufficient permissions",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiStatusBase"
            },
            "example": {
              "statusCode": 4,
              "statusMessage": "Access Denied",
              "timeStamp": 1778002000000,
              "traceId": "2dbf3d1d-06ee-4a8d-a865-1e8cc7f43a6d"
            }
          }
        }
      },
...
    },
    "parameters": {
...
      "ArticleIdPath": {
        "in": "path",
        "name": "id",
        "required": true,
        "schema": {
          "type": "integer",
          "format": "int64"
        }
      },
      "Page": {
        "in": "query",
        "name": "page",
        "schema": {
          "type": "integer",
          "default": 0,
          "minimum": 0
        }
      },
      "Size": {
        "in": "query",
        "name": "size",
        "schema": {
          "type": "integer",
          "default": 10,
          "minimum": 1,
          "maximum": 200
        }
      },
      "Sort": {
        "in": "query",
        "name": "sort",
        "schema": {
          "type": "string",
          "example": "publishedTime,desc"
        }
      },
...
    },
    "schemas": {
      "UserRole": {
        "type": "string",
        "enum": [
          "USER",
          "ADMIN"
        ]
      },
      "ArticleContentType": {
        "type": "string",
        "enum": [
          "PLAIN_TEXT",
          "MARKDOWN",
          "HTML"
        ]
      },
...
      "UserAccessResponse": {
        "type": "object",
        "example": {
          "id": 1001,
          "username": "demo_user",
          "avatarUrl": "/api/media/files/avatar_1001.png",
          "role": "USER",
          "createTime": "2026-05-05 10:00:00",
          "accessToken": "<access-token>",
          "refreshToken": "<refresh-token>"
        },
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "createTime": {
            "type": "string",
            "format": "date-time"
          },
          "accessToken": {
            "type": "string",
            "nullable": true
          },
          "refreshToken": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "ArticleMetaResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "preview": {
            "type": "string"
          },
          "contentType": {
            "$ref": "#/components/schemas/ArticleContentType"
          },
          "authorId": {
            "type": "integer",
            "format": "int64"
          },
          "publishedTime": {
            "type": "string",
            "format": "date-time"
          },
          "updatedTime": {
            "type": "string",
            "format": "date-time"
          },
          "viewCount": {
            "type": "integer"
          },
          "tags": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          }
        }
      },
      "CommentRequest": {
        "type": "object",
        "required": [
          "articleId",
          "content"
        ],
        "properties": {
          "articleId": {
            "type": "integer",
            "format": "int64"
          },
          "content": {
            "type": "string",
            "maxLength": 2000
          },
          "parentCommentId": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          }
        }
      },
      "CommentResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "articleId": {
            "type": "integer",
            "format": "int64"
          },
          "authorId": {
            "type": "integer",
            "format": "int64"
          },
          "parentCommentId": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "content": {
            "type": "string"
          },
          "createTime": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "integer"
          }
        }
      },
...
      "AdminArticleStatusRequest": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "integer",
            "minimum": 0,
            "maximum": 5
          }
        }
      },
      "AdminCommentStatusRequest": {
        "type": "object",
        "required": [
          "status"
        ],
        "example": {
          "status": 1
        },
        "properties": {
          "status": {
            "type": "integer",
            "minimum": 0,
            "maximum": 3
          }
        }
      },
      "PageArticleMetaResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ArticleMetaResponse"
            }
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer"
          },
          "number": {
            "type": "integer"
          },
          "size": {
            "type": "integer"
          }
        }
      },
...
      "PageCommentResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommentResponse"
            }
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer"
          },
          "number": {
            "type": "integer"
          },
          "size": {
            "type": "integer"
          }
        }
      },
...
      "ApiStatusBase": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer"
          },
          "statusMessage": {
            "type": "string"
          },
          "timeStamp": {
            "type": "integer",
            "format": "int64"
          },
          "traceId": {
            "type": "string"
          }
        }
      },
...
      "ApiStatusResponseString": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiStatusBase"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "string"
              }
            }
          }
        ]
      },
      "ApiStatusResponseLong": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiStatusBase"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "integer",
                "format": "int64"
              }
            }
          }
        ]
      },
      "ApiStatusResponseUserAccessResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiStatusBase"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/UserAccessResponse"
              }
            }
          }
        ]
      },
...
      "ApiStatusResponseArticleMetaResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiStatusBase"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/ArticleMetaResponse"
              }
            }
          }
        ]
      },
      "ApiStatusResponseCommentResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiStatusBase"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/CommentResponse"
              }
            }
          }
        ]
      },
...
      "ApiStatusResponsePageArticleMetaResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiStatusBase"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/PageArticleMetaResponse"
              }
            }
          }
        ]
      },
...
      "ApiStatusResponsePageCommentResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiStatusBase"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/PageCommentResponse"
              }
            }
          }
        ]
      },
...
    }
  }
}